# improve single-line completion perf by only fetching 1 completion