Running 5 iterations each of completing a large file in single-line completion mode:

n=1: 1372ms, 1181ms, 1298ms
n=3: 1958ms, 1769ms

(There was no significant perf difference for n=1 vs. n=3 on multi-line mode.)

## Test plan

Run a single-line completion and ensure it works.